feat: 지원서 개수 기능 구현

+++ src/main/java/com/kusitms/wannafly/query/dto/ApplicationFolderResponse.java
@@ -1,4 +1,4 @@
 package com.kusitms.wannafly.query.dto;
 
-public record ApplicationFolderResponse(Integer year, Integer count) {
+public record ApplicationFolderResponse(Integer year, Long count) {
 }
